decorator: Move self_hosting_management_endpoint wrapper to corporate.
